pyproject.toml
[tool.poetry]
name = "project_404"
version = "0.1.0"
description = "A chatbot to chat about Climate Change"
authors = ["gregthemitch <gmitchelljr@uchicago.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipython = "^8.21.0"
pytest = "^8.0.0"
inflect = "^7.0.0"
matplotlib = "^3.8.2"
ollama = "^0.1.6"
ipykernel = "^6.29.2"
pyngrok = "^7.1.2"
ngrok-api = "^0.10.0"
json-repair = "^0.9.0"
geopandas = "^0.14.3"
folium = "^0.15.1"
pygris = "^0.1.6"
openpyxl = "^3.1.2"
python-dotenv = "^0.19.1"
jenkspy = "^0.4.0"
colour = "^0.1.5"
fix-busted-json = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Solution from https://stackoverflow.com/questions/10253826/path-issue-with-pytest-importerror-no-module-named
# Adds home directory to sys.paths for modules
[tool.pytest.ini_options]
pythonpath = [
  "."
]